feat: update supabase&cron service for new scraper api
